requirements.txt
Flask==3.0.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
statsmodels==0.14.0
scipy==1.11.4
markdown==3.5.1
openai==1.2.3  # If using OpenAI API
google-generativeai==0.3.2  # If using Google Gemini API